// src/enums/BlyncLightProductsEnum.ts
export enum BlyncLightProductsEnum {
  BLYNCLIGHT = 'BLYNCLIGHT',
  BLYNCLIGHT_EMBRAVA_EMBEDDED = 'BLYNCLIGHT_EMBRAVA_EMBEDDED',
  BLYNCLIGHT_IPE = 'BLYNCLIGHT_IPE',
  BLYNCLIGHT_MINI = 'BLYNCLIGHT_MINI',
  BLYNCLIGHT_NAMEPLATE = 'BLYNCLIGHT_NAMEPLATE',
  BLYNCLIGHT_PLUS = 'BLYNCLIGHT_PLUS',
  BLYNCLIGHT_WIRELESS = 'BLYNCLIGHT_WIRELESS',
}